digikam from stable : USM camera bug fix : if THM file exist with a camera item, well remove it also when camera item is deleted. Tested with my Dynax 5D.
CCMAIL : digikam-devel@kde.org

svn path=/branches/stable/extragear/graphics/digikam/; revision=506377

+++ utilities/cameragui/umscamera.cpp
@@ -162,12 +162,17 @@ bool UMSCamera::getThumbnail(const QString& folder,
 
     QFileInfo fi(folder + "/" + itemName);
 
-    if (thumbnail.load(folder + "/" + fi.baseName() + ".thm"))
+    if (thumbnail.load(folder + "/" + fi.baseName() + ".thm"))        // Lowercase
     {
         if (!thumbnail.isNull())
            return true;
     }
- 
+    else if (thumbnail.load(folder + "/" + fi.baseName() + ".THM"))   // Uppercase
+    {
+        if (!thumbnail.isNull())
+           return true;
+    }   
+
     // In 3rd we trying to get thumbnail from RAW files using dcraw parse utility.
 
     KTempFile thumbFile(QString::null, "camerarawthumb");
@@ -266,8 +271,23 @@ bool UMSCamera::downloadItem(const QString& folder,
 bool UMSCamera::deleteItem(const QString& folder,
                            const QString& itemName)
 {
-    m_cancel = false;
+    d->cancel = false;
 
+    // Any camera provide THM (thumbnail) file with real image. We need to remove it also.
+
+    QFileInfo fi(folder + "/" + itemName);
+
+    QFileInfo thmLo(folder + "/" + fi.baseName() + ".thm");          // Lowercase
+
+    if (thmLo.exists())           
+        ::unlink(QFile::encodeName(thmLo.filePath()));
+
+    QFileInfo thmUp(folder + "/" + fi.baseName() + ".thm");          // Uppercase
+
+    if (thmUp.exists())           
+        ::unlink(QFile::encodeName(thmUp.filePath()));
+
+    // Remove the real image.
     return (::unlink(QFile::encodeName(folder + "/" + itemName)) == 0);
 }
